Store/pyproject.toml
[tool.poetry]
name = "store"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.14.0"
pyjwt = "^2.6.0"
requests = "^2.29.0"
python-dotenv = "^1.0.0"
gunicorn = "^20.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
